favorite

+++ resources/views/obiposts/show.blade.php
@@ -27,17 +27,27 @@
                     <p class="text-right">| {{ $obipost->created_at }}</p>
                 </div>
             </div>
+            <div class="mr-3">
+                <p class="text-right">
+                    <span class="badge badge-pill badge-danger">♡ {{ $count_favorite_users }}</span>
+                </p>
+            </div>
+            
             <div class="border-bottom pb-2">
                 <p class="text-break">{{ $obipost->content }}</p>
             </div>
             
             <div class="d-flex">
-                <div class="d-inline-flex btn btn-outline-warning btn-sm">気になる</div>
-                <div class="d-inline-flex btn btn-outline-danger btn-sm">いいね</div>
-                <div class="d-inline-flex ml-auto p-2">{!! link_to_route('obiposts.edit', '記事を編集する', ['id' => $obipost->id], ['class' => 'btn btn-success mr-1']) !!}</div>
-                {!! Form::model($obipost, ['route' => ['obiposts.destroy', $obipost->id], 'method' => 'delete']) !!}
-                <div class="d-inline-flex p-2">{!! Form::submit('記事を削除する', ['class' => 'btn btn-secondary mr-3']) !!}</div>
-                {!! Form::close() !!}
+                <div class="d-inline-flex p-2 btn btn-outline-warning btn-sm">気になる</div>
+                <div class="d-inline-flex p-2">
+                    @include('favorites.favorite_button')
+                </div>
+                @if(Auth::id() === $obipost->user->id) 
+                    <div class="d-inline-flex ml-auto p-2">{!! link_to_route('obiposts.edit', '記事を編集する', ['id' => $obipost->id], ['class' => 'btn btn-success mr-1']) !!}</div>
+                    {!! Form::model($obipost, ['route' => ['obiposts.destroy', $obipost->id], 'method' => 'delete']) !!}
+                    <div class="d-inline-flex p-2">{!! Form::submit('記事を削除する', ['class' => 'btn btn-secondary mr-3']) !!}</div>
+                    {!! Form::close() !!}
+                @endif
             </div>
         </div>
     </div>
